+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.5.0] - [Unreleased]
+
+### Changed
+
+#### ContactForm
+
+- ContactForm to TypeScript
+- useContactForm to TypeScript
+
 ## [0.4.1] - 2025-12-01
 
 ### Changed